+++ modules.tf
@@ -57,7 +57,7 @@ resource "azurerm_network_security_rule" "ssh_rule" {
 }
 
 module "bastion" {
-  source = "git::ssh://git@git.fr.clara.net/claranet/projects/cloud/azure/terraform/modules/bastion-vm.git?ref=AZ-236_use_linux-vm_base"
+  source = "github.com/claranet/terraform-azurerm-bastion-vm.git?ref=v3.1.0"
 
   client_name         = var.client_name
   location            = var.location